requirements/test.txt
-r base.txt
asynctest-0.13.0
flake8==3.7.9
isort==4.3.21
mypy==0.770
pytest-aiohttp==0.3.0
pytest-cov==2.7.1
pytest-env==0.6.2
pytest==5.1.2
coveralls==1.11.1
